pyproject.toml
[tool.poetry]
name = "data-science-project-v2"
version = "0.1.0"
description = ""
authors = ["Victor Manuel Marin Duque <victor.marin@prediktia.ai>"]
readme = "README.md"
packages = [{ include = "data_science_project_v2" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.3"
pyarrow = "^12.0.1"
requests = "^2.31.0"
google-cloud-vision = "^3.4.4"
google-cloud-storage = "^2.10.0"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
diskcache = "^5.6.1"
pylint = "^2.17.5"
mypy = "^1.4.1"

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.isort]
profile = "black"
py_version = 39
length_sort = true

[tool.mypy]
strict = true
pretty = true
ignore_missing_imports = true

[tool.pylint.main]
load-plugins = [
    "pylint.extensions.docparams",
    "pylint.extensions.docstyle",
    "pylint.extensions.code_style",
    "pylint.extensions.mccabe"
]

[tool.pylint.design]
max-complexity = 10

[tool.pylint.format]
max-line-length = "79"

[tool.pylint.parameter_documentation]
accept-no-param-doc = false
accept-no-raise-doc = false
accept-no-return-doc = false
accept-no-yields-doc = false
default-docstring-type = "default"
